latest update

+++ 11-Submission 2 Live Data/app/src/main/java/com/kotlin/submission2/data/repository/response/tv/detail/TvSeriesDetailItem.kt
@@ -15,7 +15,7 @@ data class TvSeriesDetailItem(
     val genres: List<TvSeriesGenres>,
 
     @field:SerializedName("popularity")
-    val popularity: Double,
+    val popularity: Float,
 
     @field:SerializedName("id")
     val id: Int,
